+++ apps/mvp/app/(website)/layout.tsx
@@ -1,18 +1,14 @@
-'use cache'
-
 import '../globals.css'
-import {
-  // cookies,
-  draftMode,
-} from 'next/headers'
+import {resolvePerspectiveFromCookies, type LivePerspective} from 'next-sanity/live'
 import {VisualEditing} from 'next-sanity/visual-editing'
+import {refresh, updateTag} from 'next/cache'
+import {cookies, draftMode} from 'next/headers'
+import {Suspense} from 'react'
 
 import {DebugStatus} from './DebugStatus'
 import {FormStatusLabel} from './FormStatus'
-import {SanityLive} from './live'
+import {Live} from './live'
 import {RefreshButton} from './RefreshButton'
-// import {resolvePerspectiveFromCookies} from 'next-sanity/experimental/live'
-// import {Suspense} from 'react'
 
 async function toggleDraftMode() {
   'use server'
@@ -26,6 +22,17 @@ async function toggleDraftMode() {
   }
 }
 
+async function SanityLive() {
+  let perspective: LivePerspective = 'published'
+  const isDraftMode = (await draftMode()).isEnabled
+  const jar = await cookies()
+  if (isDraftMode) {
+    perspective = await resolvePerspectiveFromCookies({cookies: jar})
+  }
+
+  return <Live perspective={perspective} />
+}
+
 export default async function RootLayout({children}: {children: React.ReactNode}) {
   const isDraftMode = (await draftMode()).isEnabled
   return (
@@ -33,6 +40,33 @@ export default async function RootLayout({children}: {children: React.ReactNode}
       <head />
       <body className="px-8">
         <div className="mt-8 mb-4 border p-4">
+          <p>Debug: {JSON.stringify({env: 'unknown'})}</p>
+          <form
+            action={async () => {
+              'use server'
+              updateTag('sanity:debug')
+            }}
+          >
+            <button type="submit">updateTag</button>
+          </form>
+          <form
+            action={async () => {
+              'use server'
+              refresh()
+            }}
+          >
+            <button type="submit">refresh</button>
+          </form>
+          <form
+            action={async () => {
+              'use server'
+              updateTag('sanity:debug')
+              refresh()
+            }}
+          >
+            <button type="submit">updateTag + refresh</button>
+          </form>
+
           <p>Draft mode: {isDraftMode ? 'On' : 'Off'}</p>
           {isDraftMode && <DebugStatus />}
           <form action={toggleDraftMode}>
@@ -44,7 +78,9 @@ export default async function RootLayout({children}: {children: React.ReactNode}
         {children}
         <RefreshButton />
         {isDraftMode && <VisualEditing />}
-        <SanityLive />
+        <Suspense>
+          <SanityLive />
+        </Suspense>
       </body>
     </html>
   )