+++ packages/next-sanity/src/shared/client-components/RefreshOnMount.tsx
@@ -6,7 +6,7 @@
  */
 
 import {useRouter} from 'next/navigation'
-import {useEffect, useReducer} from 'react'
+import {useEffect, useReducer, startTransition} from 'react'
 
 export default function RefreshOnMount(): null {
   const router = useRouter()
@@ -14,7 +14,7 @@ export default function RefreshOnMount(): null {
 
   useEffect(() => {
     if (!mounted) {
-      mount()
+      startTransition(() => mount());
       router.refresh()
     }
   }, [mounted, router])